docker/capstone-docker-default/requirements.txt
numpy<2.0
faiss-cpu==1.7.4
fastapi==0.110.2
uvicorn[standard]==0.29.0
python-multipart
pymupdf==1.25.5
sentence_transformers==2.6.1
transformers==4.39.3